+++ src/cassini/exceptions.py
@@ -1,0 +1,10 @@
+class PrintError(Exception):
+    pass
+
+
+class UploadError(Exception):
+    pass
+
+
+class PrintersError(Exception):
+    pass